refactor: extract playlist composables
